use commonjs imports, not ES modules
