+++ package.json
@@ -1,6 +1,5 @@
 {
   "private": true,
-  "type": "module",
   "scripts": {
     "dev": "next",
     "dev:watch": "next-remote-watch ./posts",
